requirements.txt
flask
python-dotenv
azure-cognitiveservices-vision-face